refactor: Extract project form into shared ProjectFormLive LiveComponent
Consolidate duplicated project creation/edit form logic from ProjectsLive,
CreateSessionLive, and ProjectComponents into a single LiveComponent that
handles its own validation, port management, and parent notification via
{:project_saved, project} messages. Uses Project.changeset/2 as the single
source of validation truth instead of manual params validation.
