+++ docs/plans/2026-04-14-002-refactor-extract-project-form-live-component-plan.md
@@ -1,7 +1,7 @@
 ---
 title: "refactor: Extract project form into shared LiveComponent"
 type: refactor
-status: active
+status: completed
 date: 2026-04-14
 ---
