+++ lib/destila_web/live/create_session_live.ex
@@ -42,12 +42,7 @@ defmodule DestilaWeb.CreateSessionLive do
      |> assign(:projects, Destila.Projects.list_projects())
      |> assign(:project_id, nil)
      |> assign(:project_step, :select)
-     |> assign(
-       :project_form,
-       to_form(%{"name" => "", "git_repo_url" => "", "local_folder" => "", "run_command" => ""})
-     )
      |> assign(:errors, %{})
-     |> assign(:port_definitions, [])
      |> assign(:page_title, Workflows.default_title(workflow_type))}
   end
 
@@ -109,87 +104,6 @@ defmodule DestilaWeb.CreateSessionLive do
     {:noreply, assign(socket, :project_step, :select)}
   end
 
-  def handle_event("validate_project_form", params, socket) do
-    port_definitions =
-      params
-      |> Enum.filter(fn {k, _} -> String.starts_with?(k, "port_def_") end)
-      |> Enum.sort_by(fn {k, _} -> k end)
-      |> Enum.map(fn {_, v} -> v end)
-
-    port_definitions =
-      if port_definitions == [], do: socket.assigns.port_definitions, else: port_definitions
-
-    {:noreply,
-     socket
-     |> assign(:project_form, to_form(params))
-     |> assign(:port_definitions, port_definitions)}
-  end
-
-  def handle_event("add_port", _params, socket) do
-    {:noreply, assign(socket, :port_definitions, socket.assigns.port_definitions ++ [""])}
-  end
-
-  def handle_event("remove_port", %{"index" => index}, socket) do
-    index = String.to_integer(index)
-
-    {:noreply,
-     assign(socket, :port_definitions, List.delete_at(socket.assigns.port_definitions, index))}
-  end
-
-  def handle_event("update_port", params, socket) do
-    index = String.to_integer(params["index"])
-    value = params["value"] || ""
-
-    {:noreply,
-     assign(
-       socket,
-       :port_definitions,
-       List.replace_at(socket.assigns.port_definitions, index, value)
-     )}
-  end
-
-  def handle_event("create_and_select_project", params, socket) do
-    name = String.trim(params["name"] || "")
-    git_repo_url = non_blank(params["git_repo_url"])
-    local_folder = non_blank(params["local_folder"])
-    run_command = non_blank(params["run_command"])
-    port_defs = Enum.reject(socket.assigns.port_definitions, &(&1 == ""))
-
-    errors = %{}
-    errors = if name == "", do: Map.put(errors, :name, "Name is required"), else: errors
-
-    errors =
-      if git_repo_url == nil && local_folder == nil do
-        Map.put(errors, :location, "Provide at least one")
-      else
-        errors
-      end
-
-    if errors == %{} do
-      {:ok, project} =
-        Destila.Projects.create_project(%{
-          name: name,
-          git_repo_url: git_repo_url,
-          local_folder: local_folder,
-          run_command: run_command,
-          port_definitions: port_defs
-        })
-
-      {:noreply,
-       socket
-       |> assign(:project_id, project.id)
-       |> assign(:projects, Destila.Projects.list_projects())
-       |> assign(:project_step, :select)
-       |> assign(:port_definitions, [])
-       |> assign(:errors, %{})}
-    else
-      {:noreply,
-       socket
-       |> assign(:project_form, to_form(params))
-       |> assign(:errors, errors)}
-    end
-  end
-
   # --- Start workflow ---
 
   def handle_event("start_workflow", params, socket) do
@@ -222,6 +136,17 @@ defmodule DestilaWeb.CreateSessionLive do
     else
       {:noreply, assign(socket, :errors, errors)}
     end
+  end
+
+  # --- Project form callback ---
+
+  def handle_info({:project_saved, project}, socket) do
+    {:noreply,
+     socket
+     |> assign(:project_id, project.id)
+     |> assign(:projects, Destila.Projects.list_projects())
+     |> assign(:project_step, :select)
+     |> assign(:errors, %{})}
   end
 
   # --- Render: type selection ---
@@ -369,9 +294,7 @@ defmodule DestilaWeb.CreateSessionLive do
               projects={@projects}
               selected_id={@project_id}
               step={@project_step}
-              form={@project_form}
               errors={@errors}
-              port_definitions={@port_definitions}
               target={nil}
             />
           </div>
@@ -422,10 +345,6 @@ defmodule DestilaWeb.CreateSessionLive do
 
     errors
   end
-
-  defp non_blank(nil), do: nil
-  defp non_blank(""), do: nil
-  defp non_blank(str), do: str
 
   defp input_heading(label, source_sessions) do
     if source_sessions != [] do
