+++ lib/destila_web/live/projects_live.ex
@@ -16,9 +16,6 @@ defmodule DestilaWeb.ProjectsLive do
      |> assign(:session_counts, Destila.Workflows.count_by_projects())
      |> assign(:creating, false)
      |> assign(:editing_project_id, nil)
-     |> assign(:form, new_form())
-     |> assign(:errors, %{})
-     |> assign(:port_definitions, [])
      |> assign(:delete_confirming_id, nil)}
   end
 
@@ -26,14 +23,10 @@ defmodule DestilaWeb.ProjectsLive do
     {:noreply,
      socket
      |> assign(:creating, true)
-     |> assign(:editing_project_id, nil)
-     |> assign(:form, new_form())
-     |> assign(:errors, %{})
-     |> assign(:port_definitions, [])}
+     |> assign(:editing_project_id, nil)}
   end
 
   def handle_event("cancel", _params, socket) do
-    # Re-stream affected projects so stream items re-render
     socket =
       socket
       |> maybe_restream_project(socket.assigns.editing_project_id)
@@ -41,114 +34,22 @@ defmodule DestilaWeb.ProjectsLive do
       |> assign(:creating, false)
       |> assign(:editing_project_id, nil)
       |> assign(:delete_confirming_id, nil)
-      |> assign(:errors, %{})
-      |> assign(:port_definitions, [])
 
     {:noreply, socket}
-  end
-
-  def handle_event("create_project", params, socket) do
-    case validate_project_params(params, socket.assigns.port_definitions) do
-      {:ok, attrs} ->
-        {:ok, _project} = Destila.Projects.create_project(attrs)
-
-        {:noreply,
-         socket
-         |> assign(:creating, false)
-         |> assign(:form, new_form())
-         |> assign(:errors, %{})
-         |> assign(:port_definitions, [])}
-
-      {:error, errors} ->
-        {:noreply,
-         socket
-         |> assign(:form, to_form(params))
-         |> assign(:errors, errors)}
-    end
   end
 
   def handle_event("edit_project", %{"id" => id}, socket) do
     project = Destila.Projects.get_project(id)
 
     if project do
-      form =
-        to_form(%{
-          "name" => project.name,
-          "git_repo_url" => project.git_repo_url || "",
-          "local_folder" => project.local_folder || "",
-          "run_command" => project.run_command || ""
-        })
-
       {:noreply,
        socket
        |> stream_insert(:projects, project)
        |> assign(:editing_project_id, id)
-       |> assign(:creating, false)
-       |> assign(:form, form)
-       |> assign(:errors, %{})
-       |> assign(:port_definitions, project.port_definitions)}
+       |> assign(:creating, false)}
     else
       {:noreply, socket}
     end
-  end
-
-  def handle_event("update_project", params, socket) do
-    id = socket.assigns.editing_project_id
-
-    case validate_project_params(params, socket.assigns.port_definitions) do
-      {:ok, attrs} ->
-        project = Destila.Projects.get_project!(id)
-        {:ok, _project} = Destila.Projects.update_project(project, attrs)
-
-        {:noreply,
-         socket
-         |> assign(:editing_project_id, nil)
-         |> assign(:form, new_form())
-         |> assign(:errors, %{})
-         |> assign(:port_definitions, [])}
-
-      {:error, errors} ->
-        project = Destila.Projects.get_project(id)
-
-        {:noreply,
-         socket
-         |> then(fn s -> if project, do: stream_insert(s, :projects, project), else: s end)
-         |> assign(:form, to_form(params))
-         |> assign(:errors, errors)}
-    end
-  end
-
-  def handle_event("validate_form", params, socket) do
-    port_definitions =
-      params
-      |> Enum.filter(fn {k, _} -> String.starts_with?(k, "port_def_") end)
-      |> Enum.sort_by(fn {k, _} -> k end)
-      |> Enum.map(fn {_, v} -> v end)
-
-    port_definitions =
-      if port_definitions == [], do: socket.assigns.port_definitions, else: port_definitions
-
-    {:noreply,
-     socket
-     |> assign(:form, to_form(params))
-     |> assign(:port_definitions, port_definitions)}
-  end
-
-  def handle_event("add_port", _params, socket) do
-    {:noreply, assign(socket, :port_definitions, socket.assigns.port_definitions ++ [""])}
-  end
-
-  def handle_event("remove_port", %{"index" => index}, socket) do
-    index = String.to_integer(index)
-    port_definitions = List.delete_at(socket.assigns.port_definitions, index)
-    {:noreply, assign(socket, :port_definitions, port_definitions)}
-  end
-
-  def handle_event("update_port", params, socket) do
-    index = String.to_integer(params["index"])
-    value = params["value"] || ""
-    port_definitions = List.replace_at(socket.assigns.port_definitions, index, value)
-    {:noreply, assign(socket, :port_definitions, port_definitions)}
   end
 
   def handle_event("confirm_delete", %{"id" => id}, socket) do
@@ -182,6 +83,20 @@ defmodule DestilaWeb.ProjectsLive do
              |> put_flash(:error, "Cannot delete this project while it is linked to sessions")}
         end
     end
+  end
+
+  # Project form callback
+
+  def handle_info({:project_saved, _project}, socket) do
+    projects = Destila.Projects.list_projects()
+
+    {:noreply,
+     socket
+     |> assign(:creating, false)
+     |> assign(:editing_project_id, nil)
+     |> stream(:projects, projects, reset: true)
+     |> assign(:projects_empty?, projects == [])
+     |> assign(:session_counts, Destila.Workflows.count_by_projects())}
   end
 
   # PubSub handlers
@@ -229,60 +144,6 @@ defmodule DestilaWeb.ProjectsLive do
     end
   end
 
-  defp new_form do
-    to_form(%{"name" => "", "git_repo_url" => "", "local_folder" => "", "run_command" => ""})
-  end
-
-  defp validate_project_params(params, port_definitions) do
-    name = String.trim(params["name"] || "")
-    git_repo_url = params["git_repo_url"]
-    git_repo_url = if git_repo_url && git_repo_url != "", do: git_repo_url, else: nil
-    local_folder = params["local_folder"]
-    local_folder = if local_folder && local_folder != "", do: local_folder, else: nil
-    run_command = params["run_command"]
-    run_command = if run_command && run_command != "", do: run_command, else: nil
-
-    # Filter empty port definitions
-    port_defs = Enum.reject(port_definitions, &(&1 == ""))
-
-    errors = %{}
-    errors = if name == "", do: Map.put(errors, :name, "Name is required"), else: errors
-
-    errors =
-      if git_repo_url == nil && local_folder == nil do
-        Map.put(errors, :location, "Provide at least one")
-      else
-        errors
-      end
-
-    # Validate port definitions format
-    errors =
-      Enum.reduce(port_defs, errors, fn pd, acc ->
-        if Regex.match?(~r/^[A-Z][A-Z0-9_]*$/, pd) do
-          acc
-        else
-          Map.put(
-            acc,
-            :port_definitions,
-            "#{pd} is not a valid env var name (use uppercase, e.g. PORT)"
-          )
-        end
-      end)
-
-    if errors == %{} do
-      {:ok,
-       %{
-         name: name,
-         git_repo_url: git_repo_url,
-         local_folder: local_folder,
-         run_command: run_command,
-         port_definitions: port_defs
-       }}
-    else
-      {:error, errors}
-    end
-  end
-
   defp linked_session_count(session_counts, project_id) do
     case Map.get(session_counts, project_id, 0) do
       0 -> "No sessions"
@@ -315,17 +176,15 @@ defmodule DestilaWeb.ProjectsLive do
         >
           <div class="card-body">
             <h3 class="text-sm font-semibold mb-3">New Project</h3>
-            <.project_form
-              form={@form}
-              errors={@errors}
-              submit_event="create_project"
-              submit_label="Create"
-              port_definitions={@port_definitions}
+            <.live_component
+              module={DestilaWeb.ProjectFormLive}
+              id="project-form-create"
+              mode={:create}
             >
               <button phx-click="cancel" type="button" class="btn btn-ghost btn-sm flex-1">
                 Cancel
               </button>
-            </.project_form>
+            </.live_component>
           </div>
         </div>
 
@@ -352,17 +211,16 @@ defmodule DestilaWeb.ProjectsLive do
             <%!-- Edit form --%>
             <div :if={@editing_project_id == project.id} class="card-body">
               <h3 class="text-sm font-semibold mb-3">Edit Project</h3>
-              <.project_form
-                form={@form}
-                errors={@errors}
-                submit_event="update_project"
-                submit_label="Save"
-                port_definitions={@port_definitions}
+              <.live_component
+                module={DestilaWeb.ProjectFormLive}
+                id={"project-form-#{project.id}"}
+                mode={:edit}
+                project={project}
               >
                 <button phx-click="cancel" type="button" class="btn btn-ghost btn-sm flex-1">
                   Cancel
                 </button>
-              </.project_form>
+              </.live_component>
             </div>
 
             <%!-- Display --%>
@@ -435,176 +293,6 @@ defmodule DestilaWeb.ProjectsLive do
         </div>
       </div>
     </Layouts.app>
-    """
-  end
-
-  attr :form, :any, required: true
-  attr :errors, :map, required: true
-  attr :submit_event, :string, required: true
-  attr :submit_label, :string, required: true
-  attr :port_definitions, :list, required: true
-  slot :inner_block
-
-  defp project_form(assigns) do
-    ~H"""
-    <form
-      phx-submit={@submit_event}
-      phx-change="validate_form"
-      class="space-y-3"
-      id={"project-form-#{@submit_event}"}
-      phx-hook="FocusFirstError"
-    >
-      <fieldset class="fieldset">
-        <label class="fieldset-label text-xs font-medium" for="project-name">
-          Name <span class="text-error">*</span>
-        </label>
-        <input
-          type="text"
-          id="project-name"
-          name="name"
-          value={@form["name"].value}
-          placeholder="My Project"
-          aria-invalid={@errors[:name] && "true"}
-          phx-mounted={JS.focus()}
-          class={[
-            "input input-bordered w-full input-sm",
-            @errors[:name] && "input-error"
-          ]}
-        />
-        <p :if={@errors[:name]} class="text-xs text-error mt-1">{@errors[:name]}</p>
-      </fieldset>
-
-      <div class={[
-        "rounded-lg p-3 space-y-3",
-        if(@errors[:location], do: "ring-1 ring-error/30 bg-error/5", else: "bg-base-200/50")
-      ]}>
-        <div class="flex items-center gap-2">
-          <span class="text-xs font-medium text-base-content/50">Location</span>
-          <span class="text-xs text-base-content/30">at least one required</span>
-        </div>
-
-        <fieldset class="fieldset">
-          <label class="fieldset-label text-xs font-medium" for="project-git-repo-url">
-            Git repository URL
-          </label>
-          <input
-            type="url"
-            id="project-git-repo-url"
-            name="git_repo_url"
-            value={@form["git_repo_url"].value}
-            placeholder="https://github.com/org/repo"
-            aria-invalid={@errors[:location] && "true"}
-            class={[
-              "input input-bordered w-full input-sm",
-              @errors[:location] && "input-error"
-            ]}
-          />
-        </fieldset>
-
-        <div class="flex items-center gap-3">
-          <div class="flex-1 h-px bg-base-300" />
-          <span class="text-xs text-base-content/30">or</span>
-          <div class="flex-1 h-px bg-base-300" />
-        </div>
-
-        <fieldset class="fieldset">
-          <label class="fieldset-label text-xs font-medium" for="project-local-folder">
-            Local folder
-          </label>
-          <input
-            type="text"
-            id="project-local-folder"
-            name="local_folder"
-            value={@form["local_folder"].value}
-            placeholder="/path/to/project"
-            aria-invalid={@errors[:location] && "true"}
-            class={[
-              "input input-bordered w-full input-sm",
-              @errors[:location] && "input-error"
-            ]}
-          />
-        </fieldset>
-
-        <p :if={@errors[:location]} class="text-xs text-error">{@errors[:location]}</p>
-      </div>
-
-      <div class="rounded-lg p-3 space-y-3 bg-base-200/50">
-        <div class="flex items-center gap-2">
-          <span class="text-xs font-medium text-base-content/50">Service</span>
-          <span class="text-xs text-base-content/30">optional</span>
-        </div>
-
-        <fieldset class="fieldset">
-          <label class="fieldset-label text-xs font-medium" for="project-run-command">
-            Run command
-          </label>
-          <input
-            type="text"
-            id="project-run-command"
-            name="run_command"
-            value={@form["run_command"].value}
-            placeholder="mix setup && mix phx.server"
-            class="input input-bordered w-full input-sm"
-          />
-        </fieldset>
-
-        <div>
-          <div class="flex items-center justify-between mb-2">
-            <label class="text-xs font-medium text-base-content/70">Port definitions</label>
-            <button
-              type="button"
-              phx-click="add_port"
-              class="btn btn-ghost btn-xs"
-              id="add-port-btn"
-            >
-              <.icon name="hero-plus-micro" class="size-3" /> Add port
-            </button>
-          </div>
-
-          <div :if={@port_definitions != []} class="space-y-2">
-            <div
-              :for={{pd, idx} <- Enum.with_index(@port_definitions)}
-              class="flex items-center gap-2"
-              id={"port-def-#{idx}"}
-            >
-              <input
-                type="text"
-                name={"port_def_#{idx}"}
-                value={pd}
-                placeholder="PORT"
-                phx-blur="update_port"
-                phx-value-index={idx}
-                class={[
-                  "input input-bordered w-full input-sm font-mono uppercase",
-                  @errors[:port_definitions] && "input-error"
-                ]}
-                id={"port-input-#{idx}"}
-              />
-              <button
-                type="button"
-                phx-click="remove_port"
-                phx-value-index={idx}
-                class="btn btn-ghost btn-xs text-error/60 hover:text-error"
-                id={"remove-port-#{idx}"}
-              >
-                <.icon name="hero-x-mark-micro" class="size-4" />
-              </button>
-            </div>
-          </div>
-
-          <p :if={@errors[:port_definitions]} class="text-xs text-error mt-1">
-            {@errors[:port_definitions]}
-          </p>
-        </div>
-      </div>
-
-      <div class="flex gap-2">
-        <button type="submit" class="btn btn-primary btn-sm flex-1">
-          {@submit_label}
-        </button>
-        {render_slot(@inner_block)}
-      </div>
-    </form>
     """
   end
 end
